Update README.md
# EasyFlyOne - Airline Company Simple Website
This project is a demonstration of a website for the sale of cheap air tickets! ***You can freely use the materials from this repository*** to implement your features. Here you can take both the design and the ticket purchase system (Server, and also the choice of everything on JS), the admin panel, tickets with identification by QR code. And much more...
## 💡 Functionality:

 - Choice for buying a ticket (From where, to where, departure date, return date, number of seats, seats on the plane)
 - Advertising
 - News
 - Authorization
 - Registration
 - Discord integration
 - Beautiful design
 - Admin panel
 - Tickets and history

The project was created using Laravel.